backend/requirements.txt
# Web
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.9

# Azure
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
azure-search-documents==11.6.0b4
azure-servicebus==7.12.2
openai==1.30.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Graph
SPARQLWrapper==2.0.0

azure-monitor-query==1.4.0
redis==5.0.4

# Telemetry
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0

# Dev / test
pytest==8.2.0
pytest-asyncio==0.23.6
ruff==0.4.4
